ocm: Move MockClusterServiceClient to a separate file
